requirements.txt
blinker==1.9.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
Flask==3.1.2
greenlet==3.2.4
gunicorn==23.0.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==25.0
playwright==1.55.0
pyee==13.0.0
python-dotenv==1.2.1
redis==7.1.0
requests==2.32.5
spotipy==2.25.2
typing_extensions==4.15.0
urllib3==2.6.3
WazeRouteCalculator==0.16
Werkzeug==3.1.5
